Added Snapfish::Downloader
Sometimes photos are not on the Snapfish CDN, usually the older photos.
In that case, we ask the Snapfish thumbnail API for the largest
thumbnail we can get. That also seems to be the behavior when you try
and download one of these photos from the site, too.

Also, enhanced existing extraction script plus added a new downloader
script that takes advantage of Snapfish::Downloader.

TODO: Add new README stuff, address TODOs in the scripts and the
downloader.